Moving plotting tools to justplotit.py
